Added all versions of EasyFarm
The commit contains all of the version of EasyFarm. The Zip folder
contains release of the ready to run version of the programs. The
release folder is the unzipped version. The Farmbot Builds folder
contains all of the working code for each version in "Release" and what
I am working on in "Experimental"