Added versions 1.0.5/6/7 to the repository.
Added versions 1.0.5/6/7 to the repository.

Switched bot from a if statement decision making process to a finite
state machine one.  Rest state is for /heal 'ing. Attack state is for
/attacking. Heal state is for abilities and spells to heal the user.
Travel state moves the bot from one waypoint to another.

Also changed the user interface to use more of a MVVM pattern; only
waypoint stuff is in the main windows code behind. The ViewModel class
is split up into multiple parts based on the tabs on the left of the
main window. The one for resting holds the interface stuff for the
resting tab etc.

Now using the more managable GameEngine objects to coordinate sharing
between different classes. It allows runtime access to almost all data
being used in the application.