Embedded FFACETools and Added all previous releases of EasyFarm to the repository.
I've embedded a copy of FFACETools into the project's executable now
lowering the number of dll files in the release/debug folders to one:
FFACE.dll.

I've also added all of the version of EasyFarm that were created except
for the initial ones that are laying around under a different project
name.