Reverted FSM Code, Added Private Server Debug Aid, Reverted Slider Changes
I've re-added the return statement in the FSM code in hopes it fixes
some intermittent state transition problems.

I've added a #if DEBUG statement in the battle function to allow
development on private servers. The private server still uses TP values
of 100 and not the new value of 1000.

I've reverted the changes in the weaponskill view. The health slider
will now trigger from the health range of 1 - 100 instead of 1 - 1000
which was a mistake.