Restore defaults button to Misc View and Changed way GetInstance works for FarmingTools
I've added a restore defaults settings to the misc tab so that once
users change the settings they can recover them back.

I've added a new GetInstance() method to FarmingTools that takes nothing
and expects the user to manually pass an FFACE session.

I've changed the way GetInstance(fface) works. It will now also return a
new instance when the user passes in an fface instance that is different
than the one stored internally.