Versions 1.1.0.3 release. Revamped casting systems, Now using XI-Tools for heavy lifting,
There's been a structural shift in the program. State classes are now
going to do the heavy lifting. This means more code will be implemented
in the states and less will be added to XI-Tools.

DebugSpellCasting helps test FFACE's spell casting functions which in
implementing the casting system.

There was a complete revamp of the combat system and the code was moved
into AttackState.cs

FarmingTools is now just a wrapper to hold the services provided by
XI-Tools. We'll be trying to move away from using this class but we're
keeping it for now until we can eliminate the TargetData, PlayerActions,
and PlayerData classes.