Ranged attack support and bugfix to stop running.
Added ranged attack support for pulling.  Created loop that will keep
trying to pull if offensive spells or ranged attacks are in the pull
list.  Before StartList and PullList are executed, run to max ranged
attack if ranged is present but no offensive spells are present,
otherwise run to max spell range if offensive spells are present.

Also added a bugfix to stop the bot from running forever if the bot is
stopped while travelling.