Added minimize to tray and recast on buff wear
Added files to minimize program to system tray and added options to cast
a spell when buff wears. The program now has its own program icon.

I've also added an resource file explorer as well. By using the list
command, it will print out the list of all attributes in an xml file. In
addition, if you type in a attribute's name, it will list all values for
the attribute within the file.