Remerging XI Tools back into EasyFarm
I've remerged XI-Tools back into easy farm for sake of updating. This
should make it easier for others to update EasyFarm without having to
look very far for all of the classes. Eventually, when we've determined
the best way to do things, we can abstract parts of the program back out
again for use in other programs.