Updated program to add in logic for mage casting support. You can now choose to lock on engage a target optionally. The update also includes a new view that adds another expander to the settings view.
The logic for fight started had to be changed as well. Now a fight is considered started when there are no pull moves or we've used the first round of pull moves already and the mobs engaged.

I've also removed the TargetInvalidComponent and have merged its functionality into the end component. The basically did the same job anyway except one checked for Units.Invalid to reset a new target.

The embedded files have been removed to allow unit testing; errors were being thrown when the dlls could not be found.