Update to include the new fixes to the casting system.
Caster now handles the correct casting of spells from start to finish and can determine whether a spell casted successfully or not.

A bug with parsing ablity and spell indexes was found that prevented the program from correctly determining recasts. IsActionValid now converts ability names to FFACETools enums to correctly determine index values on spells and abilities.

I've replaced the old function of MovingUnit with the new, cleaner version.