Centralized constants and Updated traveling to detect aggro
I've moved all the constants used to check if the program should check
the unit array into the Constants class. All checks against the unit
array should use this constant to determine the rate that the code
should check for units.

In addition, I've updated the pathing code to now detect aggro and
return control to the program when aggro occurs. This will keep the
player alive when he's caught on obstacles and can't react to aggro.