Changes to Route's View and View Model for recording button text and behavior.
We no longer rely on passing a button into the RoutesViewModel to change the record button's text. Now we've created a direct binding to a new field called "RecordRouteHeader" that the button's text binds on.

In addition, waypoints will only be recorded when there's a change in position and no longer when there's a change in heading.