Update to use new FFACE and FFACETool.
I've updated the repository to use the new FFACE and FFACETools dlls. These changes break aggro detection and fast stop on pause while traveling since the source code for this version of FFACETools is not available at this time from Lolwutt. The new dlls implement a new version of terrain travel I think that uses Final Fantasy's internal follow commands to route the player to his destinations which is a big improvement over keypress navigation, so the lose of functionality will be worth it.