Merged all partial config classes into one.
I've decided that we'll keep all config data in one place instead of close to where it's referenced. With all the data in one place, Config is still pretty small in size so it's not too much of a problem.

Process watcher has been modified to allow the program to continue when it finds a process that throws an error. When a process throws an error we won't present that process to the user so no problem.

New code is being developed to implement the combat system; it's still in the designing stage and TreeItemViewModel and new tests can be ignored.