Update UI to be more user friendly.
The old UI was too complicated for most users to use. We will be trying to transitioning back to the old UI while adding the advanced features in. The goal is to present the old ability view when the advanced features are not needed and the new one on request probably through an advanced settings link or context menu.