Simplified finite state machine, and improved theading.
I've removed a lot of the un-needed complexity from the FSM classes
which resembles Myrmidon's original code without some of the other
features. The multithreading for the main loop has been simplified,
removing the monitor that was previosly there. The new code makes
use of asynchronous methods and tasks. The EliteMMO dll has been
added but is not in use. Eventually, the user will be able to
select which library they would like to use for memory reading so
if fface is broken, but elitemmo is fixed, you can choose to use that
memory reading api.