Adding views / view models for following characters.
I've added the FollowView and FollowViewModel for following players.
In addition, I've changed the way view models are created by
simplifying the reflection done in the Locator class.

The required coupling of viewmodels to viewmodelbase has been
removed by introducing an IViewModel interface with which views
will be created with now. ViewModelBase should be used with view
model which need FFACE support.