Converted monitor classes to states.
Converted DeadMonitor and ZoneMonitor to states which will pause
the program on death. Zone state will not resume after zoning unlike
the zone monitor. The AppInformer class is now a EventPublisher class
and now contains SendPauseEvent and SendResumeEvent. FFACETools and
MMOElite wrappers and references have been moved to memory out of
EasyFarm. The BaseState class now contains a reference MemoryWrapper
for all states.