Refactor test to try out new strategy
I've refactored the tests to set up dependencies as members on
the test class. In the finder method for MemoryAPI, it will find
the latest changes to those objects, and will use them to build
up that MemoryAPI object. This way, tests don't have to change
when using new dependencies.

Not sure if this is a good idea, but a great chance to try it out!