Simplify action filter by swapping out composite for validator
Removed rule class and substitued it by using the Configurator pattern
borrowed from TopShelf C#. The idea is to yield return Result objects
so that you can run validation rules on all of a given object's
properties.

In addition, by yield returning the results, it makes the tests easier
to write since we do not need to worry about another condition returning
early, and stopping us from evaluating the state we really wanted to
test.