Mock IMemoryAPI instead of mocking everything that depends upon it
Added EliteAPIMock object to the project. Instead of mocking everything
that depends on IMemoryAPI. The benefit here is that we can set
properties on the Mock directly, without going through any loops and
hoops of mocking frameworks or dependency injection.

Going forward, this will make any authors code that uses IMemoryAPI
inherently testable.