Added explanations for how to run unit tests
using a fresh Visual studio 2019 enterprise I did not have xunit test runner installed at all so I would get a (!) on all the unit tests which would get faded out. This was because there was no test runner installed to be able to run the tests. I'm not sure if that is a per project setting, or if you install it once and it is good for the rest of your projects.

I added instructions on how to install the xunit test runner to the readme. Alternatively I'm not sure if the package.json file could be updated to include the test runner and then this problem would be solved for everyone who opens the solution and tries to run the tests.